add checkings before deleting

+++ backend/services/core-service/app/services/plan_session_main_service.py
@@ -3,12 +3,16 @@ from typing import List, Optional, Annotated
 from uuid import UUID
 from app.services.plan_session_service import PlanSessionService
 from app.domain.schemas.plan_session_schema import PlanSessionCreateSchema, PlanSessionResponseSchema, PlanSessionUpdateSchema
-
+from app.services.plan_purchase_service import PlanPurchaseService
 
 class PlanSessionMainService:
-    def __init__(self, service: Annotated[PlanSessionService, Depends()]):
+    def __init__(
+            self,
+            service: Annotated[PlanSessionService, Depends()],
+            purchase_service: Annotated[PlanPurchaseService, Depends()] ):
         self.service = service
-
+        self.purchase_service = purchase_service
+        
     async def create_session(self, data: PlanSessionCreateSchema) -> PlanSessionResponseSchema:
         session = await self.service.create(data)
         return PlanSessionResponseSchema.from_orm(session)
@@ -36,6 +40,9 @@ class PlanSessionMainService:
         return PlanSessionResponseSchema.from_orm(updated)
     
     async def delete_session(self, session_id: UUID) -> bool:
+        is_bought = await self.purchase_service.get_purchase_by_session_id(session_id)
+        if is_bought:
+            raise HTTPException(status_code=403, detail="Someone already paid for this")
         return await self.service.delete_session(session_id)
     
     async def get_sessions_by_trainer_id(self, trainer_id: UUID) -> List[PlanSessionResponseSchema]:
